% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configinit.R
\name{configinit}
\alias{configinit}
\title{configinit}
\usage{
configinit(val = list(), cfg = c("/etc/config/", "./config",
  "./config.yml"))
}
\arguments{
\item{val}{A list thet overwrite other configs}

\item{cfg}{An array of strings default = c("/etc/config/", "./config", "./config.yml")}
}
\description{
configinit reads all config file from /etc/config/*.yml, ./config/*.yml, ./config.yml
}
\examples{
 configinit(list(myconf="myconf"))
 message(getOption("myconf"))
 readLines(system.file("extdata", "example.yml", package="configoptions"))
 configinit(cfg=system.file("extdata", "example.yml", package="configoptions"))
 getOption("myconf")      # note that myconf still exists
 getOption("config_list")
 getOption("config_int")
 getOption("config_str")
}
