% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configoptions.R
\docType{package}
\name{configoptions}
\alias{configoptions}
\alias{configoptions-package}
\title{\pkg{configoptions} an R package read a folder with configs into options}
\description{
\code{configoptions} an R package read a folder with configs into options
default path is **/etc/config/**, ./config/**, ./config.yml
}
