% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setpwd.R
\name{setpwd}
\alias{setpwd}
\title{setpwd}
\usage{
setpwd()
}
\value{
current_pwd
}
\description{
set pwd for Rscript
}
\details{
set current working directory to the RScript path
setpwd ensure when the script is envoke by RScript, 
the working directory should be where the script locates
}
