% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configinit.R
\name{configinit}
\alias{configinit}
\title{configinit}
\usage{
configinit(val = list(), cfg = c("/etc/config", "./config",
  "./config.yml"))
}
\arguments{
\item{val}{A list thet overwrite other configs}

\item{cfg}{An array of strings default = c("/etc/config/", "./config", "./config.yml")}
}
\description{
configinit reads all config file (yaml) in the following places
**/etc/config/*.yml**, **./config/*.yml**, **./config.yml**
the places can be specified by user
}
\examples{
 configinit()

}
